# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "backend APIs of lifecycle manager service",
        "title": "lifecycle manager API service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "FedLCM team"
        },
        "version": "v1"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/federation/openfl": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Federation"
                ],
                "summary": "Create a new OpenFL federation",
                "parameters": [
                    {
                        "description": "The federation info",
                        "name": "federation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.FederationOpenFLCreationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success, the data field is the created federation's uuid",
                        "schema": {
                            "$ref": "#/definitions/api.GeneralResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation",
                        "schema": {
                            "$ref": "#/definitions/api.GeneralResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.GeneralResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/federation/openfl/director/yaml": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Federation"
                ],
                "summary": "Get OpenFL director deployment yaml",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the chart uuid",
                        "name": "chart_uuid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "the federation uuid",
                        "name": "federation_uuid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the deployment",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "namespace of the deployment",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "type of the service to be exposed 1: LoadBalancer 2: NodePort",
                        "name": "service_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password to access the Jupyter Notebook",
                        "name": "jupyter_password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "customized registry address",
                        "name": "registry",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "choose if use the customized registry config",
                        "name": "use_registry",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "choose if use the customized registry secret",
                        "name": "use_registry_secret",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "choose if enable the podSecurityPolicy",
                        "name": "enable_psp",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success, the data field is the yaml content",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.GeneralResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation",
                        "schema": {
                            "$ref": "#/definitions/api.GeneralResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.GeneralResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/federation/openfl/envoy/register": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Federation"
                ],
                "summary": "Process Envoy registration request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "federation UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The creation requests",
                        "name": "registrationRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.ParticipantOpenFLEnvoyRegistrationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success, the data field is the created director's uuid",
                        "schema": {
                            "$ref": "#/definitions/api.GeneralResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation",
                        "schema": {
                            "$ref": "#/definitions/api.GeneralResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.GeneralResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/federation/openfl/envoy/{uuid}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Federation"
                ],
                "summary": "Get specific info of OpenFL envoy, by providing the envoy uuid and token string",
                "parameters": [
                    {
                        "type": "string",
                        "description": "envoy UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "token string",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.GeneralResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.OpenFLEnvoyDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation",
                        "schema": {
                            "$ref": "#/definitions/api.GeneralResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.GeneralResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.GeneralResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "service.FederationOpenFLCreationRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "shard_descriptor_config": {
                    "$ref": "#/definitions/valueobject.ShardDescriptorConfig"
                },
                "use_customized_shard_descriptor": {
                    "type": "boolean"
                }
            }
        },
...
        "valueobject.ShardDescriptorConfig": {
            "type": "object",
            "properties": {
                "envoy_config_yaml": {
                    "type": "string"
                },
                "python_files": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "sample_shape": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "target_shape": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}